Added XBB virus
WHO's Technical Advisory Group on SARS-CoV-2 Virus Evolution (TAG-VE) met on the 24 October 2022 to discuss the latest evidence on the Omicron variant of concern, and how its evolution is currently unfolding, in light of high levels of population immunity in many settings and country differences in the immune landscape. In particular, the public health implications of the rise of some Omicron variants, specifically XBB and its sublineages (indicated as XBB*), as well as BQ.1 and its sublineages (indicated as BQ.1*), were discussed.